Rename orchard_flavor to flavor
