+++ src/flavor.rs
