+++ src/pczt/prover.rs
@@ -6,8 +6,8 @@ use rand::{CryptoRng, RngCore};
 use crate::{
     builder::SpendInfo,
     circuit::{Circuit, Instance, ProvingKey, Witnesses},
+    flavor::OrchardVanilla,
     note::{AssetBase, Rho},
-    orchard_flavor::OrchardVanilla,
     Note, Proof,
 };
