+++ src/pczt/tx_extractor.rs
@@ -4,7 +4,7 @@ use rand::{CryptoRng, RngCore};
 use super::Action;
 use crate::{
     bundle::{Authorization, Authorized, EffectsOnly},
-    orchard_flavor::OrchardVanilla,
+    flavor::OrchardVanilla,
     primitives::redpallas::{self, Binding, SpendAuth},
     sighash_versioning::{OrchardSighashVersion, VerBindingSig, VerSpendAuthSig},
     Proof,
